Refactor class
